well i guess its time to try and learn something
quadtrees. the VERY BASIC definition being they split an area into 4 regions(hence the name) and would allow me to check collision in those specific regions as opposed to in the entire area constantly. i believe thats the dumbed down jist of it but actually implementing them is going to be difficult for me